+++ carts/bench/conf.lua
@@ -1,4 +1,4 @@
 function love.conf(t)
-    t.console = true
+    t.console = false
     t.window = false
 end
