+++ carts/game/conf.lua
@@ -2,7 +2,7 @@ function love.conf(t)
     t.identity = "bsp"                -- The name of the save directory (string)
     t.appendidentity = false            -- Search files in source directory before save directory (boolean)
     t.version = "11.3"                  -- The LÖVE version this game was made for (string)
-    t.console = true                    -- Attach a console (boolean, Windows only)
+    t.console = false                    -- Attach a console (boolean, Windows only)
     t.window.title = "!picotron"     -- The window title (string)
     t.window.vsync = 1                  -- Vertical sync mode (number)
     t.window.msaa = 0                   -- The number of samples to use with multi-sampled antialiasing (number)
