--- carts/game/engine/game_state.lua
@@ -1,19 +1,0 @@
-local GameState={}
-local logging=require("logging")
-local input=require("engine.input")
-
--- handles console commands
-function GameState:command(cmd,...)
-end
-
-function GameState:pre_update()
-    input:update()
-end
-
-function GameState:update()
-end
-
-function GameState:draw()
-end
-
-return GameState
