--- carts/game/picotron/emulator/input.lua
@@ -1,6 +1,0 @@
-local Input={}
-local lk=love.input
-function Input.isScancodeDown(k)
-    return lk.isScancodeDown(k)
-end
-return Input
