+++ carts/game/systems/rampstyles.lua
@@ -1,0 +1,6 @@
+local RampStylesSystem={}
+local ramp_styles={}
+function RampStylesSystem:set(id, ramp)
+    ramp_styles[id] = ramp
+end
+return RampStylesSystem
